requirements.txt
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.13.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.13.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.13.0
Flask==3.0.0
psycopg2-binary==2.9.9
validators==0.22.0
redis==6.3.0